// types/web3-contracts/TransactionRequestCore.d.ts
/* Generated by ts-generator ver. 0.0.8 */
/* tslint:disable */

import Contract, { CustomOptions, contractOptions } from 'web3/eth/contract';
import { TransactionObject, BlockType } from 'web3/eth/types';
import { Callback, EventLog } from 'web3/types';
import { EventEmitter } from 'events';
import { Provider } from 'web3/providers';

export interface TransactionRequestCoreRawData {
  0: (string)[];
  1: (boolean)[];
  2: (string)[];
  3: (string)[];
}

export class TransactionRequestCore {
  constructor(jsonInterface: any[], address?: string, options?: CustomOptions);
  _address: string;
  options: contractOptions;
  methods: {
    requestData(): TransactionObject<TransactionRequestCoreRawData>;

    initialize(
      addressArgs: (string)[],
      uintArgs: (number | string)[],
      callData: (string | number[])[]
    ): TransactionObject<void>;

    execute(): TransactionObject<boolean>;

    cancel(): TransactionObject<boolean>;

    claim(): TransactionObject<boolean>;

    proxy(_to: string, _data: (string | number[])[]): TransactionObject<boolean>;

    refundClaimDeposit(): TransactionObject<boolean>;

    sendFee(): TransactionObject<boolean>;

    sendBounty(): TransactionObject<boolean>;

    sendOwnerEther(recipient: string): TransactionObject<boolean>;

    callData(): TransactionObject<(string)[]>;
  };
  deploy(options: { data: string; arguments: any[] }): TransactionObject<Contract>;
  events: {
    Aborted(
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ): EventEmitter;

    Cancelled(
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ): EventEmitter;

    Claimed(
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ): EventEmitter;

    Executed(
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ): EventEmitter;

    allEvents: (
      options?: {
        filter?: object;
        fromBlock?: BlockType;
        topics?: (null | string)[];
      },
      cb?: Callback<EventLog>
    ) => EventEmitter;
  };
  getPastEvents(
    event: string,
    options?: {
      filter?: object;
      fromBlock?: BlockType;
      toBlock?: BlockType;
      topics?: (null | string)[];
    },
    cb?: Callback<EventLog[]>
  ): Promise<EventLog[]>;
  setProvider(provider: Provider): void;
}